Switched work to an sbyte-based type temporarily
I was hitting a wall with multiplication for Fix64, so I decided to work
out the algorithms on the smallest possible base type, sbyte. With only
256 possible values, it's possible to write fast tests that cover
literally all cases. Most new work in this update is hence for Fix8.
Fix8 is not meant to be practical for use.

Features currently implemented for Fix8: addition, substraction,
multiplication, addition, abs, ceil, equality, floor, round, sign,
conversions to and from decimal, overrides of System.Object virtual
methods, IEquatable and IComparable.